+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.12.1 ()
 ## New featues
 * *Dead code* finder in puppetresources.
+* CPU related facts
 ## Bugs fixed
 * Fixed dependencies so that builds with GHC 7.8-rc1 work.